Refactor: TodoViewController 에서 TodoTableViewDelegate 분리
delegate 객체에서 dataSource 의 데이터에 접근하기위해 LinkedDataSource 프로토콜을 구현해 delegate 객체를 생성할 때 참조하도록 해줌
