+++ iOS/TodoApp/TodoApp/Todo/TodoTableViewDataSource.swift
@@ -48,7 +48,9 @@ class TodoTableViewDataSource: NSObject, UITableViewDataSource {
     func addCard(_ card: Card) {
         self.cards?.insert(card, at: 0)
     }
-    
+}
+
+extension TodoTableViewDataSource: LinkedDataSource {
     func removeCard(at index: Int) {
         self.cards?.remove(at: index)
     }
